split swift adapter layers
