+++ internal/lang/swift/types.go
@@ -1,0 +1,81 @@
+package swift
+
+import "github.com/ben-ranford/lopper/internal/lang/shared"
+
+const (
+	swiftAdapterID          = "swift"
+	packageManifestName     = "Package.swift"
+	packageResolvedName     = "Package.resolved"
+	maxDetectFiles          = 2048
+	maxScanFiles            = 4096
+	maxScannableSwiftFile   = 2 * 1024 * 1024
+	maxManifestDeclarations = 512
+	maxWarningSamples       = 5
+	ambiguousDependencyKey  = "\x00"
+)
+
+type importBinding = shared.ImportRecord
+
+type fileScan struct {
+	Path    string
+	Imports []importBinding
+	Usage   map[string]int
+}
+
+type dependencyMeta struct {
+	Declared bool
+	Resolved bool
+	Version  string
+	Revision string
+	Source   string
+}
+
+type dependencyCatalog struct {
+	Dependencies       map[string]dependencyMeta
+	AliasToDependency  map[string]string
+	ModuleToDependency map[string]string
+	LocalModules       map[string]struct{}
+}
+
+type scanResult struct {
+	Files                []fileScan
+	Warnings             []string
+	KnownDependencies    map[string]struct{}
+	ImportedDependencies map[string]struct{}
+}
+
+type repoScanner struct {
+	repoPath          string
+	catalog           dependencyCatalog
+	scan              scanResult
+	unresolvedImports map[string]int
+	foundSwift        bool
+	skippedLargeFiles int
+	visited           int
+}
+
+type swiftStringScanState struct {
+	inString     bool
+	multiline    bool
+	rawHashCount int
+	escaped      bool
+}
+
+type resolvedPin struct {
+	Identity      string `json:"identity"`
+	Package       string `json:"package"`
+	Location      string `json:"location"`
+	RepositoryURL string `json:"repositoryURL"`
+	State         struct {
+		Version  string `json:"version"`
+		Revision string `json:"revision"`
+		Branch   string `json:"branch"`
+	} `json:"state"`
+}
+
+type resolvedDocument struct {
+	Pins   []resolvedPin `json:"pins"`
+	Object struct {
+		Pins []resolvedPin `json:"pins"`
+	} `json:"object"`
+}
